# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rampart REST API
  version: 0.3.0
  summary: Uptime monitoring, status pages, incidents, and alerting.
  description: |
    Hand-curated OpenAPI description of the **primary** Rampart REST surface.
    It is representative, not exhaustive: every resource group is covered with
    its main routes and request/response shapes, but not every query parameter
    or every one of the ~100 handlers is enumerated. The canonical contract for
    any endpoint is the Rust source in
    `backend/crates/rampart-api/src/routes/` and the types in
    `backend/crates/rampart-core/`.

    ## Authentication

    Two interchangeable credentials resolve to the same authenticated user:

    * **Session cookie** — `rampart_session`, HTTP-only, `SameSite=Strict`,
      `Secure` over HTTPS. Minted by `POST /v1/auth/login` (or `/register` on
      first run). This is what the dashboard SPA uses.
    * **API key** — `Authorization: Bearer rmp_<32 chars>`, created at
      `POST /v1/api-keys`. Equivalent to a session for all `/v1` routes.

    The public surface (`/v1/auth/*`, `/v1/public/*`, the ingest receivers,
    `/push/{token}`, and the spec + health endpoints) needs no credential.

    ## RBAC

    Three roles gate the protected `/v1` surface:

    * `admin` — everything, including the admin-only groups below.
    * `editor` — full monitoring CRUD (monitors, status pages, incidents,
      notifications, tags, folders, maintenance) but no admin surfaces.
...

...
servers:
...
paths:
...
  /v1/notifications:
    get:
      tags: [Notifications]
      summary: List notification channels
      responses:
        '200':
          description: Channels.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
    post:
      tags: [Notifications]
      summary: Create a notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewNotification' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
  /v1/notifications/{id}:
    parameters:
      - $ref: '#/components/parameters/NotificationId'
    get:
      tags: [Notifications]
      summary: Get a channel
      responses:
        '200':
          description: The channel.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Notifications]
      summary: Update a channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewNotification' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
    delete:
      tags: [Notifications]
      summary: Delete a channel
      responses:
        '204': { description: Deleted. }
  /v1/notifications/{id}/test:
    parameters:
      - $ref: '#/components/parameters/NotificationId'
    post:
      tags: [Notifications]
      summary: Send a test through this channel
      responses:
        '200': { description: Sent (or per-channel error). }
  /v1/notification-templates:
    get:
      tags: [Notifications]
      summary: List message templates
      responses:
        '200':
          description: Templates.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NotificationTemplate' }
    post:
      tags: [Notifications]
      summary: Create a message template
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationTemplate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationTemplate' }
  /v1/notification-templates/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Notifications]
      summary: Get a template
      responses:
        '200':
          description: The template.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationTemplate' }
    patch:
      tags: [Notifications]
      summary: Update a template
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationTemplate' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationTemplate' }
    delete:
      tags: [Notifications]
      summary: Delete a template
      responses:
        '204': { description: Deleted. }

  # ─────────────────────────── Tags ───────────────────────────
...
  /v1/monitors/{id}/notifications:
    parameters:
      - $ref: '#/components/parameters/MonitorId'
    get:
      tags: [Notifications]
      summary: List notification channels attached to a monitor
      responses:
        '200': { description: Attached channel ids. }
  /v1/monitors/{id}/notifications/{nid}:
    parameters:
      - $ref: '#/components/parameters/MonitorId'
      - name: nid
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Notifications]
      summary: Attach a notification channel to a monitor
      responses:
        '204': { description: Attached. }
    delete:
      tags: [Notifications]
      summary: Detach a notification channel from a monitor
      responses:
        '204': { description: Detached. }
...
  /v1/notification-templates/preview:
    post:
      tags: [Notifications]
      summary: Render a template against sample data
      responses:
        '200': { description: Rendered preview. }
        '400': { $ref: '#/components/responses/BadRequest' }

  # ─────────────────────────── Folders (extended: tags + channels) ───────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rampart_session
      description: HTTP-only session cookie set by `/v1/auth/login`.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: rmp_<32 chars>
      description: Personal API key created via `/v1/api-keys`.
  parameters:
    MonitorId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    NotificationId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      description: Error envelope produced by `ApiError`.
...
    ChannelKind:
      type: string
      description: Notification channel kind. Serialized snake_case. ~130 supported; a representative subset is listed; `custom` is the catch-all webhook.
      enum:
        - slack
        - discord
        - telegram
        - email
        - webhook
        - pagerduty
        - sms_twilio
        - teams
        - gotify
        - pushover
        - ntfy
        - mattermost
        - matrix
        - signal
        - rocket_chat
        - google_chat
        - apprise
        - custom
        - webpush
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { $ref: '#/components/schemas/ChannelKind' }
        config: { type: object, description: 'Kind-specific config. Secrets encrypted before insert.' }
        is_default: { type: boolean }
        template_id: { type: [string, 'null'], format: uuid }
        active: { type: boolean }
        last_fired_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        cooldown_seconds: { type: integer }
      required: [id, name, kind, config, is_default, active, created_at]
    NewNotification:
      type: object
      required: [name, kind, config]
      properties:
        name: { type: string }
        kind: { $ref: '#/components/schemas/ChannelKind' }
        config: { type: object }
        is_default: { type: boolean }
        template_id: { type: [string, 'null'], format: uuid }
        active: { type: boolean }
        cooldown_seconds: { type: integer }
    NotificationTemplate:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        channel_kinds: { type: array, items: { type: string } }
        event_kind: { type: string }
        subject_template: { type: [string, 'null'] }
        body_template: { type: string }
        is_default: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [name, channel_kinds, event_kind, body_template]
...
security:
  - sessionCookie: []
  - bearerApiKey: []